requirements.txt
numpy<2.0.0
ultralytics==8.1.0
opencv-python==4.9.0.80
pydantic==2.6.1
pyyaml==6.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6